Fix errors in Vigenere module.
The 'vigenereBy' function would hang when given an empty key string, so
this is now checked by a guard. With an empty key it simply returns the
plaintext. The underlying issue is that 'chunksOf' returns an infinite
list of empty lists for chunks of 0, which makes sense. We can't divide
the text that way, so a reasonable default is null encryption.

Incorrect output would be produced when the length of the original raw
key did not match the length of the standardized key. Using the former
in the local 'sections' function produces sections that are too long and
will not zip correctly. The final output was garbled and shorter than
the correct output. This has been fixed by using the length of the key
that is actually used for encryption.

The local 'round' function was renamed so as not to clash with the
mathematical Prelude function.